Struct kimchi_circuits::gate::CircuitGate

pub struct CircuitGate<F: FftField> {
    pub typ: GateType,
    pub wires: GateWires,
    pub c: Vec<F>,
}

Fields
typ: GateType

type of the gate

wires: GateWires

gate wires

c: Vec<F>

constraints vector

Implementations

impl<F: FftField> CircuitGate<F>

pub fn zero(wires: GateWires) -> Self

this function creates “empty” circuit gate

pub fn verify(
&self,
row: usize,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

This function verifies the consistency of the wire assignements (witness) against the constraints

impl<F: FftField> CircuitGate<F>

pub fn create_endomul(wires: GateWires) -> Self

pub fn verify_endomul(
&self,
row: usize,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

pub fn endomul(&self) -> F

impl<F: FftField> CircuitGate<F>

pub fn create_generic(wires: GateWires, qw: [F; 3], qm: F, qc: F) -> Self

pub fn create_generic_add(
wires: GateWires,
left_coeff: F,
right_coeff: F
) -> Self

creates an addition gate

pub fn create_generic_mul(wires: GateWires) -> Self

creates a multiplication gate

pub fn create_generic_const(wires: GateWires, constant: F) -> Self

creates a constant gate

pub fn create_generic_public(wires: GateWires) -> Self

creates a public input gate

pub fn verify_generic(
&self,
row: usize,
witness: &[Vec<F>; 15]
) -> Result<(), String>

verifies that the generic gate constraints are solved by the witness

pub fn generic(&self) -> F

impl<F: FftField> CircuitGate<F>

pub fn create_poseidon(wires: GateWires, c: [[F; 3]; 5]) -> Self

pub fn create_poseidon_gadget(
row: usize,
first_and_last_row: [GateWires; 2],
round_constants: &[Vec<F>]
) -> (Vec<Self>, usize)

create_poseidon_gadget(row, first_and_last_row, round_constants) creates an entire set of constraint for a Poseidon hash. For that, you need to pass:

the index of the first row

the first and last rows’ wires (because they are used in the permutation)

the round constants The function returns a set of gates, as well as the next pointer to the circuit (next empty absolute row)

pub fn verify_poseidon(
&self,
row: usize,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

Checks if a witness verifies a poseidon gate

pub fn ps(&self) -> F

pub fn rc(&self) -> [[F; 3]; 5]

round constant that are relevant for this specific gate

impl<F: FftField> CircuitGate<F>

pub fn create_vbmul(wires: &[GateWires; 2]) -> Vec<Self>

pub fn verify_vbmul(
&self,
_row: usize,
_witness: &[Vec<F>; 15]
) -> Result<(), String>

pub fn vbmul(&self) -> F

impl<F: FftField> CircuitGate<F>

pub fn verify_complete_add(
&self,
row: usize,
witness: &[Vec<F>; 15]
) -> Result<(), String>

Check the correctness of witness values for a complete-add gate.

impl<F: FftField> CircuitGate<F>

pub fn verify_endomul_scalar(
&self,
row: usize,
witness: &[Vec<F>; 15],
_cs: &ConstraintSystem<F>
) -> Result<(), String>