Module kimchi_circuits::nolookup::constraints

Re-exports

pub use crate::polynomial::WitnessEvals;

pub use crate::polynomial::WitnessOverDomains;

pub use crate::polynomial::WitnessShifts;

Structs

ConstraintSystem

Shifts

Shifts represent the shifts required in the permutation argument of PLONK. It also caches the shifted powers of omega for optimization purposes.

Enums

GateError

Represents an error found when verifying a witness with a gate

Constants

ZK_ROWS

Functions

eval_vanishes_on_last_4_rows

Evaluates the polynomial (x - w^{n - 4}) (x - w^{n - 3}) * (x - w^{n - 2}) * (x - w^{n - 1})

eval_zk_polynomial

Evaluates the polynomial (x - w^{n - 3}) * (x - w^{n - 2}) * (x - w^{n - 1})

vanishes_on_last_4_rows

The polynomial (x - w^{n - 4}) (x - w^{n - 3}) * (x - w^{n - 2}) * (x - w^{n - 1})

zk_polynomial

Computes the zero-knowledge polynomial for blinding the permutation polynomial: (x-w^{n-k})(x-w^{n-k-1})...(x-w^n). Currently, we use k = 3 for 2 blinding factors, see https://www.plonk.cafe/t/noob-questions-plonk-paper/73

zk_w3

Returns the end of the circuit, which is used for introducing zero-knowledge in the permutation polynomial