Struct kimchi_circuits::nolookup::constraints::ConstraintSystem

pub struct ConstraintSystem<F: FftField> {

    pub public: usize,
    pub domain: EvaluationDomains<F>,
    pub gates: Vec<CircuitGate<F>>,
    pub sigmam: [DP<F>; 7],
    pub zkpm: DP<F>,
    pub coefficientsm: [DP<F>; 15],
    pub coefficients8: [E<F, D<F>>; 15],
    pub genericm: DP<F>,
    pub psm: DP<F>,
    pub complete_addm: DP<F>,
    pub mulm: DP<F>,
    pub emulm: DP<F>,
    pub endomul_scalarm: DP<F>,
    pub generic4: E<F, D<F>>,
    pub sigmal1: [E<F, D<F>>; 7],
    pub sigmal8: [E<F, D<F>>; 7],
    pub sid: Vec<F>,
    pub ps8: E<F, D<F>>,
    pub complete_addl4: E<F, D<F>>,
    pub mull8: E<F, D<F>>,
    pub emull: E<F, D<F>>,
    pub chacha8: Option<[E<F, D<F>>; 4]>,
    pub endomul_scalar8: E<F, D<F>>,
    pub l1: E<F, D<F>>,
    pub l04: E<F, D<F>>,
    pub l08: E<F, D<F>>,
    pub zero4: E<F, D<F>>,
    pub zero8: E<F, D<F>>,
    pub zkpl: E<F, D<F>>,
    pub vanishes_on_last_4_rows: E<F, D<F>>,
    pub shift: [F; 7],
    pub endo: F,
    pub fr_sponge_params: ArithmeticSpongeParams<F>,
    pub dummy_lookup_values: Vec<Vec<F>>,
    pub lookup_tables: Vec<Vec<DP<F>>>,
    pub lookup_tables8: Vec<Vec<E<F, D<F>>>>,
    pub lookup_table_lengths: Vec<usize>,
    pub lookup_selectors: Vec<E<F, D<F>>>,

}

Fields
public: usize

number of public inputs

domain: EvaluationDomains<F>

evaluation domains

gates: Vec<CircuitGate<F>>

circuit gates

sigmam: [DP<F>; 7]

permutation polynomial array

zkpm: DP<F>

zero-knowledge polynomial

coefficientsm: [DP<F>; 15]

coefficients polynomials in coefficient form

coefficients8: [E<F, D<F>>; 15]

coefficients polynomials in evaluation form

genericm: DP<F>psm: DP<F>

poseidon constraint selector polynomial

complete_addm: DP<F>

EC point addition constraint selector polynomial

mulm: DP<F>

mulm constraint selector polynomial

emulm: DP<F>

emulm constraint selector polynomial

endomul_scalarm: DP<F>

endomul scalar computation

generic4: E<F, D<F>>

multiplication evaluations over domain.d4

sigmal1: [E<F, D<F>>; 7]

permutation polynomial array evaluations over domain d1

sigmal8: [E<F, D<F>>; 7]

permutation polynomial array evaluations over domain d8

sid: Vec<F>

SID polynomial

ps8: E<F, D<F>>

poseidon selector over domain.d8

complete_addl4: E<F, D<F>>

EC point addition selector evaluations w over domain.d4

mull8: E<F, D<F>>

scalar multiplication selector evaluations over domain.d8

emull: E<F, D<F>>

endoscalar multiplication selector evaluations over domain.d8

chacha8: Option<[E<F, D<F>>; 4]>

ChaCha indexes

endomul_scalar8: E<F, D<F>>

EC point addition selector evaluations w over domain.d8

l1: E<F, D<F>>

1-st Lagrange evaluated over domain.d8

l04: E<F, D<F>>

0-th Lagrange evaluated over domain.d4

l08: E<F, D<F>>

0-th Lagrange evaluated over domain.d8

zero4: E<F, D<F>>

zero evaluated over domain.d8

zero8: E<F, D<F>>

zero evaluated over domain.d8

zkpl: E<F, D<F>>

zero-knowledge polynomial over domain.d8

vanishes_on_last_4_rows: E<F, D<F>>

the polynomial that vanishes on the last four rows

shift: [F; 7]

wire coordinate shifts

endo: F

coefficient for the group endomorphism

fr_sponge_params: ArithmeticSpongeParams<F>

random oracle argument parameters

dummy_lookup_values: Vec<Vec<F>>

Lookup tables

lookup_tables: Vec<Vec<DP<F>>>lookup_tables8: Vec<Vec<E<F, D<F>>>>lookup_table_lengths: Vec<usize>lookup_selectors: Vec<E<F, D<F>>>

Lookup selectors: For each kind of lookup-pattern, we have a selector that’s 1 at the rows where that pattern should be enforced, and 1 at all other rows.

Implementations

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn create(
gates: Vec<CircuitGate<F>>,
lookup_tables: Vec<Vec<Vec<F>>>,
fr_sponge_params: ArithmeticSpongeParams<F>,
public: usize
) -> Option<Self>

creates a constraint system from a vector of gates (CircuitGate), some sponge parameters (ArithmeticSpongeParams), and the number of public inputs.

pub fn verify(&self, witness: &[Vec<F>; 15]) -> Result<(), GateError>

This function verifies the consistency of the wire assignements (witness) against the constraints witness: wire assignement witness RETURN: verification status

pub fn evaluate(&self, w: &[DP<F>; 15], z: &DP<F>) -> WitnessOverDomains<F>

evaluate witness polynomials over domains

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn gnrc_quot(
&self,
witness_d4: &[Evaluations<F, D<F>>; 15]
) -> Evaluations<F, D<F>>

generic constraint quotient poly contribution computation

pub fn gnrc_scalars(w_zeta: &[F; 15], generic_zeta: F) -> Vec<F>

produces generic(zeta) * w0 * w1, generic(zeta) * w0, generic(zeta) * w1, generic(zeta) * w2

pub fn gnrc_lnrz(&self, w_zeta: &[F; 15], generic_zeta: F) -> DensePolynomial<F>

generic constraint linearization poly contribution computation

pub fn verify_generic(
&self,
witness: &[DensePolynomial<F>; 15],
public: &DensePolynomial<F>
) -> bool

function to verify the generic polynomials with a witness

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn perm_quot(
&self,
lagrange: &WitnessOverDomains<F>,
beta: F,
gamma: F,
z: &DensePolynomial<F>,
alpha: &[F]
) -> Result<(Evaluations<F, D<F>>, DensePolynomial<F>), ProofError>

permutation quotient poly contribution computation

pub fn perm_lnrz(
&self,
e: &[ProofEvaluations<F>],
zeta: F,
beta: F,
gamma: F,
alpha: &[F]
) -> DensePolynomial<F>

permutation linearization poly contribution computation

pub fn perm_scalars(
e: &[ProofEvaluations<F>],
beta: F,
gamma: F,
alpha: &[F],
zkp_zeta: F
) -> F

pub fn perm_aggreg(
&self,
witness: &[Vec<F>; 15],
beta: &F,
gamma: &F,
rng: &mut impl RngCore + CryptoRng
) -> Result<DensePolynomial<F>, ProofError>

permutation aggregation polynomial computation

impl<F: FftField + SquareRootField> ConstraintSystem<F>

pub fn psdn_scalars(
evals: &[ProofEvaluations<F>],
params: &ArithmeticSpongeParams<F>,
alpha: &[F]
) -> Vec<F>

pub fn psdn_lnrz(
&self,
evals: &[ProofEvaluations<F>],
params: &ArithmeticSpongeParams<F>,
alpha: &[F]
) -> DensePolynomial<F>

poseidon linearization poly contribution computation f^7 + c(x) - f(wx)