Module kimchi_circuits::polynomials::lookup

This source file implements the arithmetization of plookup constraints

Because of our ZK-rows, we can’t do the trick in the plookup paper of wrapping around to enforce consistency between the sorted lookup columns.

Instead, we arrange the LookupSorted table into columns in a snake-shape.

Like so, _ _ | | | | | | | | | | || || |

or, imagining the full sorted array is [ s0, …, s8 ], like

s0 s4 s4 s8 s1 s3 s5 s7 s2 s2 s6 s6

So the direction (“increasing” or “decreasing” (relative to LookupTable) is if i % 2 = 0 { Increasing } else { Decreasing }

Then, for each i < max_lookups_per_row, if i % 2 = 0, we enforce that the last element of LookupSorted(i) = last element of LookupSorted(i + 1), and if i % 2 = 1, we enforce that the first element of LookupSorted(i) = first element of LookupSorted(i + 1)

Structs

CombinedEntry

UncombinedEntry

Constants

ZK_ROWS

The number of random values to append to columns for zero-knowledge.

Traits

Entry

Functions

aggregation

Computes the aggregation polynomial for maximum n lookups per row, whose kth entry is the product of terms

constraints

Specifies the lookup constraints as expressions.

sorted

Computes the sorted lookup tables required by the lookup argument.

verify

Checks that all the lookup constraints are satisfied.

zk_patch

Pad with zeroes and then add 3 random elements in the last two rows for zero knowledge.